+++ src/api/BdmsContextExtensions.cs
@@ -87,10 +87,7 @@ public static class BdmsContextExtensions
         List<int> drillingMudTypeIds = codelists.Where(c => c.Schema == "drilling_mud_type").Select(s => s.Id).ToList();
         List<int> identifierIds = codelists.Where(c => c.Schema == "borehole_identifier").Select(s => s.Id).ToList();
 
-        // Lithology codelists (schema without LithologySchemas static class belong to legacy schemas)
-        List<int> chronostratigraphyTopBedrockIds = codelists.Where(c => c.Schema == "custom.chronostratigraphy_top_bedrock").Select(s => s.Id).ToList();
-        List<int> lithostratigraphyTopBedrockIds = codelists.Where(c => c.Schema == "custom.lithostratigraphy_top_bedrock").Select(s => s.Id).ToList();
-
+        // Lithology codelists
         List<int> plasticityIds = codelists.Where(c => c.Schema == StratigraphySchemas.PlasticitySchema).Select(s => s.Id).ToList();
         List<int> compactnessIds = codelists.Where(c => c.Schema == StratigraphySchemas.CompactnessSchema).Select(s => s.Id).ToList();
         List<int> consistencyIds = codelists.Where(c => c.Schema == StratigraphySchemas.ConsistencySchema).Select(s => s.Id).ToList();
@@ -1252,7 +1249,7 @@ public static class BdmsContextExtensions
             .StrictMode(true)
             .RuleFor(o => o.FromDepth, f => (chronostratigraphy_ids % 10) * 10)
             .RuleFor(o => o.ToDepth, f => ((chronostratigraphy_ids % 10) + 1) * 10)
-            .RuleFor(o => o.ChronostratigraphyId, f => f.PickRandom(chronostratigraphyTopBedrockIds).OrNull(f, .05f))
+            .RuleFor(o => o.ChronostratigraphyId, f => f.PickRandom(chronostratigraphyIds).OrNull(f, .05f))
             .RuleFor(o => o.Chronostratigraphy, _ => default!)
             .RuleFor(o => o.StratigraphyId, f => GetStratigraphyId(chronostratigraphy_ids, 11_000_000))
             .RuleFor(o => o.Stratigraphy, _ => default!)
@@ -1283,7 +1280,7 @@ public static class BdmsContextExtensions
             .StrictMode(true)
             .RuleFor(o => o.FromDepth, f => (lithostratigraphy_ids % 10) * 10)
             .RuleFor(o => o.ToDepth, f => ((lithostratigraphy_ids % 10) + 1) * 10)
-            .RuleFor(o => o.LithostratigraphyId, f => f.PickRandom(lithostratigraphyTopBedrockIds).OrNull(f, .05f))
+            .RuleFor(o => o.LithostratigraphyId, f => f.PickRandom(lithostratigraphyIds).OrNull(f, .05f))
             .RuleFor(o => o.Lithostratigraphy, _ => default!)
             .RuleFor(o => o.StratigraphyId, f => GetStratigraphyId(lithostratigraphy_ids, 14_000_000))
             .RuleFor(o => o.Stratigraphy, _ => default!)